service
add serviceAuth and auto add folder for each created folder in file name

+++ server/src/controller/files.controller.ts
@@ -1,5 +1,5 @@
 import type { RequestHandler } from 'express'
-import { writeFile } from 'fs'
+import { existsSync, writeFile } from 'fs'
 import { createFileSchema } from '../schema/files.schema'
 import { prisma } from '../db/prisma'
 import { createFolder, decodeBase64ToFile } from '../util/customhelper.js'
@@ -26,6 +26,7 @@ export const createFile: RequestHandler = async (req, res) => {
     fileUserId: req.context.user!.userId,
     fileMimeType: req.body.fileMimeType,
     fileName: req.body.fileName,
+    filePath: req.body.filePath,
   })
 
   if (!validatedBody.success) {
@@ -36,28 +37,46 @@ export const createFile: RequestHandler = async (req, res) => {
     const file = Buffer.from(validatedBody.data.file, 'base64')
 
     await prisma.$transaction(async (prisma) => {
-      const filePath = `./root/${validatedBody.data.fileName}`
+      const filePath = `./root${validatedBody.data.filePath}/${validatedBody.data.fileName}`
 
-      writeFile(filePath, file, (err) => {
-        if (err) throw err
-        console.log('The file has been saved!')
-      })
+      if (existsSync(filePath)) {
+        // make revision entry of file
+        return res.status(400).json({ error: 'File already exists!' })
+      } else {
+        writeFile(filePath, file, (err) => {
+          if (err) throw err
+          console.log('The file has been saved!')
+        })
 
-      const newFileId = nanoid()
-      const newFile = await prisma.files.create({
-        data: {
-          fileBase: validatedBody.data.file,
-          fileId: newFileId,
-          fileName: validatedBody.data.fileName,
-          filePath,
-          fileUserId: req.context.user!.userId,
-        },
-        omit: {
-          fileBase: true,
-        },
-      })
+        const newFileId = nanoid()
+        const newFile = await prisma.files.create({
+          data: {
+            fileBase: validatedBody.data.file,
+            fileId: newFileId,
+            fileName: validatedBody.data.fileName,
+            filePath: validatedBody.data.filePath,
+            fileUserId: req.context.user!.userId,
+          },
+          omit: {
+            fileBase: true,
+          },
+        })
+        return res.status(200).json({ data: newFile })
+      }
 
-      return res.status(200).json({ data: newFile })
+      // const folder = await prisma.folders.findMany({
+      //   where: {
+      //     folderName: validatedBody.data.fileName.split('_')[0],
+      //   },
+      // })
+
+      // await prisma.folderFiles.create({
+      //   data: {
+      //     ffId: newFileId,
+      //     ffFolderId: folderId,
+      //     ffFileId: newFileId,
+      //   },
+      // })
     })
   } catch (err) {
     return res.status(500).json({ error: err })
@@ -68,6 +87,7 @@ export const uploadFile: RequestHandler = async (req, res) => {
   try {
     const { filename, filecontent } = req.body
     const targetFolder = `${CONFIG.IS_FILE_SERVER == true ? CONFIG.FILE_SERVER : 'root'}/data`
+    const userId = req.context.user!.userId
     let folders = filename.split('_')
 
     if (filename.includes('REV')) {
@@ -81,33 +101,31 @@ export const uploadFile: RequestHandler = async (req, res) => {
       let departmentFolder = `${yearOrArchive}/${folders[2]}`
       let documentTypeFolder = `${departmentFolder}/${folders[3]}`
 
-      createFolder(targetFolder)
-      createFolder(companyFolder)
-      createFolder(yearOrArchive)
-      createFolder(departmentFolder)
-      createFolder(documentTypeFolder)
+      await createFolder(targetFolder, 0, userId)
+      await createFolder(companyFolder, 1, userId)
+      await createFolder(yearOrArchive, 2, userId)
+      await createFolder(departmentFolder, 3, userId)
+      await createFolder(documentTypeFolder, 4, userId)
 
       decodeBase64ToFile(filecontent, `${documentTypeFolder}/${filename}`)
 
       await prisma.$transaction(async (prisma) => {
         const newFileId = nanoid()
-        const newFile = await prisma.files.create({
+        await prisma.files.create({
           data: {
             fileBase: filecontent,
             fileId: newFileId,
             fileName: filename,
             filePath: documentTypeFolder,
-            fileUserId: 'U6MDlGyfG0fx1h1i09eeV',
+            fileUserId: userId,
           },
           omit: {
             fileBase: true,
           },
         })
-  
-       
       })
     } else {
-      console.log('Incorrect Fileame!', filename)
+      console.log('Incorrect Filename!', filename)
     }
 
     res.status(200).send({ msg: 'Upload Success' })
